fixed admin dashboard and changed file names
