+++ editAdoption.php
@@ -23,7 +23,6 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
     $adoption->gender = htmlspecialchars(trim($_POST['gender']));
     $adoption->contact = htmlspecialchars(trim($_POST['contact']));
     $adoption->monthly_salary = htmlspecialchars(trim($_POST['monthly_salary']));
-    $adoption->pet_type = htmlspecialchars(trim($_POST['pet_type']));
 
     // Validate animal_id
     $animalData = $animal->getAnimalById($animal_id);
@@ -138,16 +137,6 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         
         <label for="monthly_salary">Monthly Salary:</label>
         <input type="number" id="monthly_salary" name="monthly_salary" value="<?php echo htmlspecialchars($adoptionData['monthly_salary']); ?>" required><br><br>
-        
-        <label for="pet_type">Select Pet Type:</label><br>
-        <select id="pet_type" name="pet_type" required>
-            <option value="">--Select--</option>
-            <option value="Cat" <?php echo ($adoptionData['pet_type'] == 'Cat') ? 'selected' : ''; ?>>Cat</option>
-            <option value="Dog" <?php echo ($adoptionData['pet_type'] == 'Dog') ? 'selected' : ''; ?>>Dog</option>
-            <option value="Rabbit" <?php echo ($adoptionData['pet_type'] == 'Rabbit') ? 'selected' : ''; ?>>Rabbit</option>
-            <option value="Guinea Pig" <?php echo ($adoptionData['pet_type'] == 'Guinea Pig') ? 'selected' : ''; ?>>Guinea Pig</option>
-            <option value="Hamster" <?php echo ($adoptionData['pet_type'] == 'Hamster') ? 'selected' : ''; ?>>Hamster</option>
-        </select><br><br>
         
         <button type="submit">Update Adoption</button>
     </form>
