+++ fetchAnimalData.php
@@ -1,13 +1,13 @@
 <?php
 require_once 'classes/dbconnection.php';
-require_once 'classes/crudAnimal.php';
+require_once 'classes/IncidentManagement.php';
 
 // Initialize database connection
 $db = new Database();
 $conn = $db->getConnect();
 
 // Instantiate the Animals class with the database connection
-$crudAnimal = new Animals($conn);
+$crudIncident = new Incident($conn);
 
 // Fetch statistics for animals
 $totalInFacility = $crudAnimal->getTotalInFacility();
